Move clients into subdirectory for cleaner package structure
