+++ src/alma_api_client/__init__.py
@@ -1,5 +1,5 @@
-from .alma_api_client import AlmaAPIClient
-from .alma_analytics_client import AlmaAnalyticsClient
+from .clients.alma_api_client import AlmaAPIClient
+from .clients.alma_analytics_client import AlmaAnalyticsClient
 from .models.api import APIResponse
 from .models.marc_records import AuthorityRecord, BibRecord, HoldingRecord
 from .models.sets import Set
